requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
cssselect2==0.6.0
Django==3.2.6
django-categories==1.8
django-environ==0.8.1
django-js-asset==2.0.0
django-mptt==0.11.0
django-phonenumber-field==6.1.0
django-seed==0.3.1
django-taggit==3.0.0
easy-thumbnails==2.8.1
emoji==1.7.0
Faker==13.12.0
gunicorn==20.1.0
idna==3.3
lxml==4.8.0
phonenumbers==8.12.49
Pillow==9.1.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
reportlab==3.6.9
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
toposort==1.7
unicode-slugify==0.1.3
Unidecode==1.3.4
urllib3==1.26.9
webencodings==0.5.1
